Move artifact operation lookup to queryAst
